requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-braces==1.14.0
django-chunked-upload==2.0.0
django-cors-headers==3.2.1
django-ftpserver==0.7.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-social-auth==0.7.28
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-chunked-upload==0.4.2
drf-yasg==1.17.1
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httpie==2.0.0
httplib2==0.17.2
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.3
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyftpdlib==1.5.6
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
